+++ cmd/optimize/status.go
@@ -15,10 +15,6 @@
 package optimize
 
 import (
-	"fmt"
-	"net/url"
-	"strings"
-
 	"github.com/apex/log"
 	"github.com/spf13/cobra"
 
@@ -32,7 +28,6 @@ func init() {
 }
 
 func NewCmdStatus() *cobra.Command {
-	flags := commonFlags{}
 	statusCmd := &cobra.Command{
 		Use:   "status",
 		Short: "List onboarded optimizer configuration and status",
@@ -45,23 +40,23 @@ You may also specify a particular optimizer ID to fetch details for a single opt
 `,
 		Example:          "fsoc optimize status --workload-name frontend",
 		Args:             cobra.NoArgs,
-		RunE:             listStatus(&flags),
+		RunE:             listStatus,
 		TraverseChildren: true,
 		Annotations: map[string]string{
 			output.TableFieldsAnnotation:  "OPTIMIZERID: .id, WORKLOADNAME: .data.optimizer.target.k8sDeployment.workloadName, STATUS: .data.optimizerState, SUSPENDED: .data.suspended, STAGE: .data.optimizationState, AGENT: .data.agentState, TUNING: .data.tuningState, BLOCKERS: (.data.optimizer.ignoredBlockers? // \"false\" | select(. == \"false\") // \"true\")",
 			output.DetailFieldsAnnotation: "OPTIMIZERID: .id, CONTAINER: .data.optimizer.target.k8sDeployment.containerName, WORKLOADNAME: .data.optimizer.target.k8sDeployment.workloadName, NAMESPACE: .data.optimizer.target.k8sDeployment.namespaceName, CLUSTER: .data.optimizer.target.k8sDeployment.clusterName, STATUS: .data.optimizerState, SUSPENDED: .data.suspended, SUSPENSIONS: .data.optimizer.suspensions, RESTARTEDAT: .data.optimizer.restartTimestamp, STAGE: .data.optimizationState, AGENT: .data.agentState, TUNING: .data.tuningState, BLOCKERS: (.data.optimizer.ignoredBlockers?.blockers? // {} | keys)",
 		},
 	}
-	statusCmd.Flags().StringVarP(&flags.Cluster, "cluster", "c", "", "Filter statuses by kubernetes cluster name")
-	statusCmd.Flags().StringVarP(&flags.Namespace, "namespace", "n", "", "Filter statuses by kubernetes namespace")
-	statusCmd.Flags().StringVarP(&flags.WorkloadName, "workload-name", "w", "", "Filter statuses by name of kubernetes workload")
+	statusCmd.Flags().StringP("cluster", "c", "", "Filter statuses by kubernetes cluster name")
+	statusCmd.Flags().StringP("namespace", "n", "", "Filter statuses by kubernetes namespace")
+	statusCmd.Flags().StringP("workload-name", "w", "", "Filter statuses by name of kubernetes workload")
 
-	statusCmd.Flags().StringVarP(&flags.optimizerId, "optimizer-id", "i", "", "Retrieve status for a specific optimizer by its ID (best used with -o detail)")
+	statusCmd.Flags().StringP("optimizer-id", "i", "", "Retrieve status for a specific optimizer by its ID (best used with -o detail)")
 	statusCmd.MarkFlagsMutuallyExclusive("optimizer-id", "cluster")
 	statusCmd.MarkFlagsMutuallyExclusive("optimizer-id", "namespace")
 	statusCmd.MarkFlagsMutuallyExclusive("optimizer-id", "workload-name")
 
-	statusCmd.Flags().StringVarP(&flags.solutionName, "solution-name", "", "optimize", "Intended for developer usage, overrides the name of the solution defining the Orion types for reading/writing")
+	statusCmd.Flags().StringP("solution-name", "", "optimize", "Intended for developer usage, overrides the name of the solution defining the Orion types for reading/writing")
 	if err := statusCmd.LocalFlags().MarkHidden("solution-name"); err != nil {
 		log.Warnf("Failed to set statusCmd solution-name flag hidden: %v", err)
 	}
@@ -74,47 +69,10 @@ You may also specify a particular optimizer ID to fetch details for a single opt
 	return statusCmd
 }
 
-func listStatus(flags *commonFlags) func(cmd *cobra.Command, args []string) error {
-	return func(cmd *cobra.Command, args []string) error {
-		objStoreUrl := getKnowledgeURL(cmd, "status", "data.optimizer")
+func listStatus(cmd *cobra.Command, args []string) error {
+	objStoreUrl := getKnowledgeURL(cmd, "status", "data.optimizer")
 
-		headers := getOrionTenantHeaders()
-		cmdkit.FetchAndPrint(cmd, objStoreUrl, &cmdkit.FetchAndPrintOptions{Headers: headers, IsCollection: true})
-		return nil
-	}
-}
-
-func getKnowledgeURL(cmd *cobra.Command, objName string, objectPathPrefix string) string {
-	solutionName := cmd.Flag("solution-name").Value.String()
-	objStoreUrl := fmt.Sprintf("knowledge-store/v1/objects/%v:%s", solutionName, objName)
-
-	filterSegments := make([]string, 0, 4)
-	flags := cmd.Flags()
-	if flags != nil {
-		var val string
-		val, _ = flags.GetString("optimizer-id")
-		if val != "" {
-			filterSegments = append(filterSegments, fmt.Sprintf("id eq %q", val))
-		}
-		val, _ = flags.GetString("cluster")
-		if val != "" {
-			filterSegments = append(filterSegments, fmt.Sprintf("%s.target.k8sDeployment.clusterName eq %q", objectPathPrefix, val))
-		}
-		val, _ = flags.GetString("namespace")
-		if val != "" {
-			filterSegments = append(filterSegments, fmt.Sprintf("%s.target.k8sDeployment.namespaceName eq %q", objectPathPrefix, val))
-		}
-		val, _ = flags.GetString("workload-name")
-		if val != "" {
-			filterSegments = append(filterSegments, fmt.Sprintf("%s.target.k8sDeployment.workloadName eq %q", objectPathPrefix, val))
-		}
-	}
-
-	filterCriteria := strings.Join(filterSegments, " and ")
-	if filterCriteria != "" {
-		query := fmt.Sprintf("filter=%s", url.QueryEscape(filterCriteria))
-		objStoreUrl = objStoreUrl + "?" + query
-	}
-
-	return objStoreUrl
+	headers := getOrionTenantHeaders()
+	cmdkit.FetchAndPrint(cmd, objStoreUrl, &cmdkit.FetchAndPrintOptions{Headers: headers, IsCollection: true})
+	return nil
 }
